Move ClientOutcome to data_types.
